Replace and update Talkies dialog system
Removed the old demo/talkies.lua and added an updated version in libs/talkies.lua with improved features and customization options. Updated main.lua to use the new Talkies module and set additional color and style properties for dialog appearance.
